Locator ftw
